moved components into separate module with CommonModule
