+++ src/app/components/app/app.component.css
