+++ src/main.ts
@@ -1,7 +1,7 @@
 import { enableProdMode } from '@angular/core';
 import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
 
-import { TableModule } from './app/app.module';
+import { TableModule } from './app/table.module';
 import { environment } from './environments/environment';
 
 if (environment.production) {
